feat: implement adding of new story
Added the HTML form for inputting info for a new story and the "submit"
link in the navigation bar to display the form.

In nav.js, added the JavaScript for displaying the initially hidden new
story form when clicking on the "submit" link in the navigation bar.

In stories.js, added the logic for taking the input values from the
form, passing it to the method that will contact the API, and
displaying the new story in the stories list.

In models.js, implemented the logic for calling the API to post a new
story.  Also added the new story to the stories array of StoryList for
future use and to be consistent with what is displayed on the webpage.
The new story had to be placed in the front of the array to maintain
consistent ordering.  Also added the story to the current user's own
stories, since that won't be updated until the webpage reloads.