feat: allow favoriting and show favorites list
In the HTML, added the favorites link in the navigation bar, added an
ordered list element for putting the list of favorites, and updated the
Font Awesome link version to be able to use up-to-date icons.

In nav.js, added a function to create and show the list of favorites.

In stories.js, added in a star when creating a story list item, if
there is a user logged in.  Prepending the star to the common string of
HTML elements, along with this logic of assigning the star icon, will
make it easier to add additional icons.  Also, the current
implementation of going thru the list of favorites instead of using
something else, like a Set, to find which story was favorited, was
because it was simpler to implement, in consideration to the already
existing starter code.

In stories.js, the putStoriesOnPage function was refactored/repurposed
to work for displaying the favorites, and any other lists, as well.
All previous uses of putStoriesOnPage were refactored.

In user.js, added the handling of when a favorite-star is clicked.

In models.js, added the logic for contacting the API to add and remove
a story from the current user's favorites.  This also updates the local
favorites array within the current user instance, so that the favorites
list is up-to-date.