fix: move and hide login forms
The forms for logging in and signing up were beings placed at the
bottom of the webpage.  It is now moved to the top, like how it is in
the working demo.

The forms were still being shown after logging in.  updateNavOnLogin
was updated to recreate the UI, and its name was changed accordingly.