feat: add section for current user's own stories
Added a link in the navigation bar that will go to the current user's
own story section.  In other words, display a list of stories that the
current user has created.

Added the ordered list HTML element for the current user's own stories
in the HTML document, in order to have a place to append the stories.

In main.js, added the appropriate constant variables for reducing the
need to re-find HTML elements.

In nav.js, added a function to hide other webpage components and show
the own stories component, when clicking on the "my stories" navigation
link.