feat: allow deleting of story
In stories.js, added a trash icon for a story when generating a list
item HTML element for it, to allow it to be deleted.  The choice to
scan through a user's ownStories array, in order to determine if the
story belongs to the current user and whether to show the trash icon,
for every story to be added to lists was to simplify the amount of code
to be written.  In the future, a data structure can be used to improve
performance, by having it store the story IDs, and then checking if a
story ID exists in the data structure.

In stories.js, added a function to handle the deletion and removal of
the story whenever a trash icon is clicked.

In models.js, added a deleteStory method to the Story class.  This
seemed like the appropriate place, according to the API documentation.
This deleteStory method calls the API and deletes the story locally.
Deleting the story locally is required, since story data is not fetched
unless the webpage is reloaded.  With this methodology, there will be a
local version that can get out of sync with the online database
version, but this methodology will reduce API calls.