fix: change DOM traversal for finding parent
Using the function parentElement to find an element's story-parent was
unreliable.  If additional elements were added to the "li" parent, then
parentElement could give incorrect results.

Instead of parentElement, the function "closest" was used to find the
closest "li" parent, which would always be the top HTML element for a
story and would always contain the story ID.