feat: update CSS
Modified CSS to match the designed mock-up.

Added in "div" elements to the inside of story "li" elements, during
creation, to better organize and select the components within the "li".

Added in classes for the icons and text of a story "li" element, and
updated the insertion of the star and trash icons to utilize the new
classes.

Added in horizontal rules where needed.