feat: add error handling for log in and sign up
As per the instructions, only added error handling for logging in and
signing up.  Other possible places for handling were skipped to save
time.  The handling was done in user.js, instead of within the classes
in models.js, because displaying error messages was more appropriate
within user.js, since API-related code within models.js is supposed to
be solely for contacting the API and relaying responses.