feat: add ability to edit own story
In stories.js, added a handler, for when a pencil icon is clicked, to
populate and reuse the submit form with the associated story's info.
Since the pencil icon is only applied to the current user's created
stories, editing someone else's story is not possible.

In stories.js, modified the submitStory function to also handle
updating an edited story, which allowed for reusing the same code.  The
associated story "li" elements will be replaced with new elements,
because any one of author, title, or URL could have been changed, and
it allowed reusing an existing function (generateStoryMarkup) that
already provided what was needed.

In stories.js, a global variable, storyToEdit, was used, because it was
the best solution to using the other functions that required a Story
instance, it was the easiest way that allowed the submitStory function
to also handle updating an edited story, and it reduces the need to
find the Story instance again after new info has been submitted.  A
check of if storyToEdit is truthy is much simpler than creating a
separate and nearly exact form just to handle editing or adding classes
to the submit form to determine if it is submitting a new story or
editing an existing one.

In models.js, added a method that updated the current user's story in
the database and in local stories arrays.  It was placed in the Story
class to align with how the API has its story-updating endpoint.  It
was made as an instance method to reduce code, because if it were
static, it would have to take in extra arguments for new story info and
Story instance, have extra code to re-find Story instances, or put the
logic of updating Story instances out of the models.js file and into an
inappropriate location.

In models.js, also added a helper method to find a Story in an array of
Stories and update it.

In main.js, created global variables for the submit form's inputs.