fix & refactor: hiding and displaying stories
In main.js -> hidePageComponents(), removed the individual all-stories
and favorites list and replaced with the JQuery object, $storiesList,
which contains all of the stories lists.

In stories.js -> putStoriesOnPage(stories, list), removed the hiding of
stories lists, because it was redundant.  Before calling
putStoriesOnPage, the page components would have already been hidden.

In stories.js -> submitStory(evt), prepending a new story to one or all
of the all-stories, favorites, or own-stories HTML lists caused
unnecessary complexity and unexpected behavior.  It was simpler and
provided more consistent behavior to navigate to, recreate, and
redisplay a particular HTML list.  As a result, generating the HTML was
no longer needed.